(GH-34) Fix - commands do not interact with desktop
Commands are not executed interactively because they are run as part of
a scheduled task (by vagrant), so there is no chance for windows to pop
open that scripts like AHK (auto hot key) is waiting to kill and move
on. Change it so that vagrant runs the scripts interactively so that
script executors like AHK can kill windows and not time out the
installation.