(GH-29) Existing package reverification
Every 30 days all existing packages need to be retested. This continues
to ensure that the latest and greatest version of a package continues
to work appropriately and give folks a heads up when their packages
need to be updated. Granted, at 30 days it may be a bit past when other
folks are telling them the package is broken. This may be moved up to
every 15 days at some point.