(GH-96) Check for presence of vmId file
Check for the presence of the vmId file once at the start of each test
execution.  If not present, then don't attempt to change the TEMP folder
location, or attempt image capture.  This should allow correct startup
of VM after a complete refresh of the system, including removal of VM.